# Add new transitive_parameters method to accelerate library loading