--- a/.github/workflows/cd.yml
+++ b/.github/workflows/cd.yml
@@ -23,7 +23,7 @@ jobs:
         with:
           ref: ${{github.event.pull_request.head.ref}}
       - name: setup-python
-        uses: actions/setup-python@v4
+        uses: actions/setup-python@v5
         with:
           python-version: '3.11'
       # # update repo visualization for docs
@@ -66,7 +66,7 @@ jobs:
       - name: checkout
         uses: actions/checkout@v4
       - name: setup-python
-        uses: actions/setup-python@v4
+        uses: actions/setup-python@v5
       # install poetry and build dist
       - name: install-poetry
         uses: snok/install-poetry@v1