--- a/buildingmotif/dataclasses/template.py
+++ b/buildingmotif/dataclasses/template.py
@@ -4,7 +4,6 @@ import warnings
 from collections import Counter
 from copy import copy
 from dataclasses import dataclass
-from functools import cached_property
 from io import BytesIO, StringIO
 from itertools import chain
 from os import PathLike
@@ -140,7 +139,7 @@ class Template:
         """
         self._bm.table_connection.delete_template_dependency(self.id, dependency.id)
 
-    @cached_property
+    @property
     def all_parameters(self) -> Set[str]:
         """The set of all parameters used in this template *including* its
         dependencies. Includes optional parameters.
@@ -156,7 +155,7 @@ class Template:
             params.update(dep.template.parameters)
         return params
 
-    @cached_property
+    @property
     def parameters(self) -> Set[str]:
         """The set of all parameters used in this template *excluding* its
         dependencies. Includes optional parameters.
@@ -169,7 +168,7 @@ class Template:
         params = {str(p)[len(PARAM) :] for p in nodes if str(p).startswith(PARAM)}
         return params
 
-    @cached_property
+    @property
     def dependency_parameters(self) -> Set[str]:
         """The set of all parameters used in this template's dependencies, including
         optional parameters.
@@ -182,7 +181,7 @@ class Template:
             params = params.union(dep.template.parameters)
         return params
 
-    @cached_property
+    @property
     def parameter_counts(self) -> Counter:
         """An addressable histogram of the parameter name counts in this
         template and all of its transitive dependencies.
@@ -239,7 +238,7 @@ class Template:
         replace_nodes(templ.body, to_replace)
         return templ
 
-    @cached_property
+    @property
     def transitive_parameters(self) -> Set[str]:
         """Get all parameters used in this template and its dependencies.