--- a/tests/integration/fixtures/bacnet/Dockerfile
+++ b/tests/integration/fixtures/bacnet/Dockerfile
@@ -9,7 +9,7 @@ RUN apt update \
     && rm -rf /var/lib/apt/lists/*
 
 COPY requirements.txt .
-RUN pip3 install -r requirements.txt
+RUN python3 -m pip install -r requirements.txt
 
 COPY virtual_bacnet.py virtual_bacnet.py
 COPY BACpypes.ini .